[QA] restructure rpc tests directory
* move non-test classes to subdir `test-framework`
