+++ qa/rpc-tests/bipdersig.py
@@ -7,9 +7,8 @@
 # Test the BIP66 changeover logic
 #
 
-from test_framework import BitcoinTestFramework
-from bitcoinrpc.authproxy import AuthServiceProxy, JSONRPCException
-from util import *
+from test_framework.test_framework import BitcoinTestFramework
+from test_framework.util import *
 import os
 import shutil
