+++ qa/rpc-tests/invalidateblock.py
@@ -7,9 +7,8 @@
 # Test InvalidateBlock code
 #
 
-from test_framework import BitcoinTestFramework
-from bitcoinrpc.authproxy import AuthServiceProxy, JSONRPCException
-from util import *
+from test_framework.test_framework import BitcoinTestFramework
+from test_framework.util import *
 
 class InvalidateTest(BitcoinTestFramework):
