+++ qa/rpc-tests/keypool.py
@@ -8,7 +8,7 @@
 # Add python-bitcoinrpc to module search path:
 import os
 import sys
-sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)), "python-bitcoinrpc"))
+sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)), "test_framework/python-bitcoinrpc"))
 
 import json
 import shutil
@@ -17,7 +17,7 @@ import tempfile
 import traceback
 
 from bitcoinrpc.authproxy import AuthServiceProxy, JSONRPCException
-from util import *
+from test_framework.util import *
 
 
 def check_array_result(object_array, to_match, expected):
