+++ qa/rpc-tests/maxblocksinflight.py
@@ -4,9 +4,9 @@
 # file COPYING or http://www.opensource.org/licenses/mit-license.php.
 #
 
-from mininode import *
-from test_framework import BitcoinTestFramework
-from util import *
+from test_framework.mininode import *
+from test_framework.test_framework import BitcoinTestFramework
+from test_framework.util import *
 import logging
 
 '''
