+++ qa/rpc-tests/mempool_coinbase_spends.py
@@ -8,9 +8,8 @@
 # that spend (directly or indirectly) coinbase transactions.
 #
 
-from test_framework import BitcoinTestFramework
-from bitcoinrpc.authproxy import AuthServiceProxy, JSONRPCException
-from util import *
+from test_framework.test_framework import BitcoinTestFramework
+from test_framework.util import *
 import os
 import shutil
