+++ qa/rpc-tests/mempool_spendcoinbase.py
@@ -13,9 +13,8 @@
 # but less mature coinbase spends are NOT.
 #
 
-from test_framework import BitcoinTestFramework
-from bitcoinrpc.authproxy import AuthServiceProxy, JSONRPCException
-from util import *
+from test_framework.test_framework import BitcoinTestFramework
+from test_framework.util import *
 import os
 import shutil
