+++ qa/rpc-tests/pruning.py
@@ -11,9 +11,8 @@
 # This test takes 30 mins or more (up to 2 hours)
 # ********
 
-from test_framework import BitcoinTestFramework
-from bitcoinrpc.authproxy import AuthServiceProxy, JSONRPCException
-from util import *
+from test_framework.test_framework import BitcoinTestFramework
+from test_framework.util import *
 import os.path
 
 def calc_usage(blockdir):
