+++ qa/rpc-tests/reindex.py
@@ -6,9 +6,8 @@
 #
 # Test -reindex with CheckBlockIndex
 #
-from test_framework import BitcoinTestFramework
-from bitcoinrpc.authproxy import AuthServiceProxy, JSONRPCException
-from util import *
+from test_framework.test_framework import BitcoinTestFramework
+from test_framework.util import *
 import os.path
 
 class ReindexTest(BitcoinTestFramework):
