+++ qa/rpc-tests/rest.py
@@ -7,8 +7,9 @@
 # Test REST interface
 #
 
-from test_framework import BitcoinTestFramework
-from util import *
+
+from test_framework.test_framework import BitcoinTestFramework
+from test_framework.util import *
 from struct import *
 import binascii
 import json
