+++ qa/rpc-tests/test_framework/comptool.py
