+++ qa/rpc-tests/test_framework/mininode.py
