+++ qa/rpc-tests/test_framework/netutil.py
