+++ qa/rpc-tests/test_framework/python-bitcoinrpc/bitcoinrpc/.gitignore
