+++ qa/rpc-tests/test_framework/socks5.py
